+++ jwt-test/jwt-junit5-spring/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,1 @@
-org.springframework.boot.env.EnvironmentPostProcessor=\
-org.entur.jwt.spring.test.JwtEnvironmentPostProcessor
+org.springframework.test.context.ContextCustomizerFactory=org.entur.jwt.spring.test.JwtTestContextCustomizerFactory
