--- jwt-test/jwt-junit5-spring/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-org.entur.jwt.spring.test.JwtEnvironmentPostProcessor
